requirements.txt
absl-py==2.1.0
ace_tools==0.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==5.0.1
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
coverage==7.6.10
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.55.3
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.0.post2
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
networkx==3.2.1
numba==0.60.0
numpy>=1.21,<2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
pooch==1.6.0
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.0
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-reapy==0.10.0
pytube==15.0.0
pytz==2024.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.11.3
six==1.17.0
soundfile==0.12.1
soxr==0.5.0.post1
spotipy==2.24.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.6.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.2
yt-dlp==2024.12.23
zipp==3.21.0
